Add reusable components and change initial view
This commit adds a few reusable components in the "components" folder
in lib, as well as redefining the home screen a little. My idea was to
have the entry screen be a list of all the stored characters (presumably
read in from a JSON array), and then when clicking on the add character
button a popup window would display all the fields and stuff for creating
a new character.

TODO: Haven't yet done the popup window, will look at this tomorrow. Just
wanted to give an idea of where I was going with this.
TODO ALSO: I kinda wanna break up that big character file into several smaller
files, so it's easier to find stuff. And maybe simplify how data is being passed
between them? To be figured out at a later date